contributions/2_contribution/pyproject.toml
[project]
name = "feature-reliance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations==1.4.1",
    "hydra-core==1.3.2",
    "matplotlib==3.9.1",
    "medmnist==3.0.2",
    "datasets==2.21.0",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "open-clip-torch>=2.32.0",
    "opencv-python-headless==4.10.0.84",
    "pandas==2.2.2",
    "pillow==10.4.0",
    "pytorch-lightning==2.3.3",
    "pyyaml==6.0.1",
    "setuptools==75.1.0",
    "scikit-image==0.24.0",
    "scikit-learn==1.7.1",
    "timm==1.0.9",
    "torch==2.4.0",
    "torchmetrics==1.6.2",
    "torchvision==0.19.0",
    "tqdm==4.66.5",
    "transformers>=5.4.0",
    "kagglehub[pandas-datasets]>=1.0.0",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
